Bug 1820252 - Remove viewport-translated r=#translations-reviewers!
This patch removes the functionality that reported to the Translations
actors when the viewport is translated. This metric was sometimes
unreliable due to dynamic content coming into the viewport, whereas
the metric itself only considered the viewport on page load.

This metric is less important with the upcomping move to intersection observer.

Differential Revision: https://phabricator.services.mozilla.com/D249971