Bug 1820252 - Update TranslationsDocument mutation tests r=#translations-reviewers!
This patch makes several updates to the TranslationsDocument
mutations tests. The mock port now captures the counts of each
kind of request that is sent to the TranslationsEngine for
assertions, and the comparisons themselves are now done with
regex, instead of via prettified strings.

This allows us to test cases in which the translated text may
lie within a range of translationId values, as long as the
expected number of requests, and the final translated output
match the expectation.

Also updates all existing test cases to be compatible with
the new scheduling algorithm, as well as adds many new cases.

Differential Revision: https://phabricator.services.mozilla.com/D249990