Bug 1975744 - Ensure that a not-confident language identifications shows the translation button; r?#translations-reviewers!
There was an error in the test case, which made the behavior incorrect.
I had to refactor the detected languages because I needed a good way to
detect if the detection ran or not. Previously you had to check the
identifiedLangConfident value as true, false, or undefined. It felt
cleaner to just pass on the detected languages from the original API.

Differential Revision: https://phabricator.services.mozilla.com/D256207