Bug 2006407 - Add a ./mach eval command that forwards arguments to perftest r=sparky!
This makes it more ergonomic for running evals as a unique test type.
Originally I wanted this to run like `./mach eval test.js -- --headless`
where everything past the -- would be forwarded to the underlying test.
However I struggled with the argument parsing in mach having surprising
behavior and swallowing the -- before the mach command was run.

Mostly this serves as an entrypoint and area for future documentation
for browser evals.

Differential Revision: https://phabricator.services.mozilla.com/D277467